fix memory leak related to expression flag, undo removing facelineColorIsTransparent
* Address potential memory leak that would happen if: you set values
past the first byte in FFLAllExpressionFlag or if the values were
uninitialized AND FFL_MODEL_FLAG_NEW_EXPRESSIONS was not set.
- This is because, while ExecuteGPUStep tries to initialize the mid
and high flags to zero when that flag is not set, the rest of the
function continues to use the version of that flag that didn't have that
zeroing applied. It uses the original FFLCharModelDesc while the zeroing
is applied to the CharModel's CharModelDesc.

* Walk back change to remove facelineColorIsTransparent property.
- Going to do this in FFL-Testing as well. Embarrassingly, I noticed
that beards were broken with the Switch shader for a WHOLE MONTH because
of this...
- ... yeah, never making that mistake again. I'm still not sure why
the color for the beard is off with this, but...

* Add more logging for FFL_RESULT_ERROR cases. Now it should ALL be
accounted for whenever you see that.

+++ include/nn/ffl/FFLiAuthorID.h
@@ -3,7 +3,7 @@
 
 #include <nn/ffl/types.h>
 
-#define FFLI_AUTHOR_ID_SIZE  (sizeof(u64))
+#define FFLI_AUTHOR_ID_SIZE  (int)sizeof(u64)
 
 typedef struct FFLiAuthorID
 {
